+++ src/Symfony/Component/Form/Form.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\Form;
 
 use Symfony\Component\Form\Exception\RuntimeException;
 use Symfony\Component\Form\Exception\UnexpectedTypeException;
-use Symfony\Component\Form\Exception\AlreadyBoundException;
+use Symfony\Component\Form\Exception\AlreadySubmittedException;
 use Symfony\Component\Form\Exception\TransformationFailedException;
 use Symfony\Component\Form\Exception\LogicException;
 use Symfony\Component\Form\Exception\OutOfBoundsException;
@@ -84,10 +84,10 @@ class Form implements \IteratorAggregate, FormInterface
     private $errors = array();
 
     /**
-     * Whether this form is bound
+     * Whether this form was submitted
      * @var Boolean
      */
-    private $bound = false;
+    private $submitted = false;
 
     /**
      * The form data in model format
@@ -108,7 +108,7 @@ class Form implements \IteratorAggregate, FormInterface
     private $viewData;
 
     /**
-     * The bound values that don't belong to any children
+     * The submitted values that don't belong to any children
      * @var array
      */
     private $extraData = array();
@@ -244,8 +244,8 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function setParent(FormInterface $parent = null)
     {
-        if ($this->bound) {
-            throw new AlreadyBoundException('You cannot set the parent of a bound form');
+        if ($this->submitted) {
+            throw new AlreadySubmittedException('You cannot set the parent of a submitted form');
         }
 
         if (null !== $parent && '' === $this->config->getName()) {
@@ -286,11 +286,11 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function setData($modelData)
     {
-        // If the form is bound while disabled, it is set to bound, but the data is not
+        // If the form is submitted while disabled, it is set to submitted, but the data is not
         // changed. In such cases (i.e. when the form is not initialized yet) don't
         // abort this method.
-        if ($this->bound && $this->defaultDataSet) {
-            throw new AlreadyBoundException('You cannot change the data of a bound form.');
+        if ($this->submitted && $this->defaultDataSet) {
+            throw new AlreadySubmittedException('You cannot change the data of a submitted form.');
         }
 
         // If the form inherits its parent's data, disallow data setting to
@@ -464,26 +464,26 @@ class Form implements \IteratorAggregate, FormInterface
     /**
      * {@inheritdoc}
      */
-    public function bind($submittedData)
+    public function submit($submittedData)
     {
-        if ($this->bound) {
-            throw new AlreadyBoundException('A form can only be bound once');
+        if ($this->submitted) {
+            throw new AlreadySubmittedException('A form can only be submitted once');
         }
 
         // Initialize errors in the very beginning so that we don't lose any
         // errors added during listeners
         $this->errors = array();
 
-        // Obviously, a disabled form should not change its data upon binding.
+        // Obviously, a disabled form should not change its data upon submission.
         if ($this->isDisabled()) {
-            $this->bound = true;
+            $this->submitted = true;
 
             return $this;
         }
 
         // The data must be initialized if it was not initialized yet.
         // This is necessary to guarantee that the *_SET_DATA listeners
-        // are always invoked before bind() takes place.
+        // are always invoked before submit() takes place.
         if (!$this->defaultDataSet) {
             $this->setData($this->config->getData());
         }
@@ -498,10 +498,10 @@ class Form implements \IteratorAggregate, FormInterface
 
         $dispatcher = $this->config->getEventDispatcher();
 
-        // Hook to change content of the data bound by the browser
-        if ($dispatcher->hasListeners(FormEvents::PRE_BIND)) {
+        // Hook to change content of the data submitted by the browser
+        if ($dispatcher->hasListeners(FormEvents::PRE_SUBMIT)) {
             $event = new FormEvent($this, $submittedData);
-            $dispatcher->dispatch(FormEvents::PRE_BIND, $event);
+            $dispatcher->dispatch(FormEvents::PRE_SUBMIT, $event);
             $submittedData = $event->getData();
         }
 
@@ -515,7 +515,7 @@ class Form implements \IteratorAggregate, FormInterface
             }
 
             foreach ($this->children as $name => $child) {
-                $child->bind(isset($submittedData[$name]) ? $submittedData[$name] : null);
+                $child->submit(isset($submittedData[$name]) ? $submittedData[$name] : null);
                 unset($submittedData[$name]);
             }
 
@@ -529,11 +529,11 @@ class Form implements \IteratorAggregate, FormInterface
         // its parent's data) into account.
         // (see InheritDataAwareIterator below)
         if ($this->config->getInheritData()) {
-            $this->bound = true;
+            $this->submitted = true;
 
-            // When POST_BIND is reached, the data is not yet updated, so pass
+            // When POST_SUBMIT is reached, the data is not yet updated, so pass
             // NULL to prevent hard-to-debug bugs.
-            $dataForPostBind = null;
+            $dataForPostSubmit = null;
         } else {
             // If the form is compound, the default data in view format
             // is reused. The data of the children is merged into this
@@ -558,7 +558,7 @@ class Form implements \IteratorAggregate, FormInterface
             if (count($this->children) > 0) {
                 // Use InheritDataAwareIterator to process children of
                 // descendants that inherit this form's data.
-                // These descendants will not be bound normally (see the check
+                // These descendants will not be submitted normally (see the check
                 // for $this->config->getInheritData() above)
                 $childrenIterator = new InheritDataAwareIterator($this->children);
                 $childrenIterator = new \RecursiveIteratorIterator($childrenIterator);
@@ -574,9 +574,9 @@ class Form implements \IteratorAggregate, FormInterface
 
                 // Hook to change content of the data into the normalized
                 // representation
-                if ($dispatcher->hasListeners(FormEvents::BIND)) {
+                if ($dispatcher->hasListeners(FormEvents::SUBMIT)) {
                     $event = new FormEvent($this, $normData);
-                    $dispatcher->dispatch(FormEvents::BIND, $event);
+                    $dispatcher->dispatch(FormEvents::SUBMIT, $event);
                     $normData = $event->getData();
                 }
 
@@ -587,20 +587,31 @@ class Form implements \IteratorAggregate, FormInterface
                 $this->synchronized = false;
             }
 
-            $this->bound = true;
+            $this->submitted = true;
             $this->modelData = $modelData;
             $this->normData = $normData;
             $this->viewData = $viewData;
 
-            $dataForPostBind = $viewData;
+            $dataForPostSubmit = $viewData;
         }
 
-        if ($dispatcher->hasListeners(FormEvents::POST_BIND)) {
-            $event = new FormEvent($this, $dataForPostBind);
-            $dispatcher->dispatch(FormEvents::POST_BIND, $event);
+        if ($dispatcher->hasListeners(FormEvents::POST_SUBMIT)) {
+            $event = new FormEvent($this, $dataForPostSubmit);
+            $dispatcher->dispatch(FormEvents::POST_SUBMIT, $event);
         }
 
         return $this;
+    }
+
+    /**
+     * Alias of {@link submit()}.
+     *
+     * @deprecated Deprecated since version 2.3, to be removed in 3.0. Use
+     *             {@link submit()} instead.
+     */
+    public function bind($submittedData)
+    {
+        return $this->submit($submittedData);
     }
 
     /**
@@ -620,9 +631,20 @@ class Form implements \IteratorAggregate, FormInterface
     /**
      * {@inheritdoc}
      */
+    public function isSubmitted()
+    {
+        return $this->submitted;
+    }
+
+    /**
+     * Alias of {@link isSubmitted()}.
+     *
+     * @deprecated Deprecated since version 2.3, to be removed in 3.0. Use
+     *             {@link isSubmitted()} instead.
+     */
     public function isBound()
     {
-        return $this->bound;
+        return $this->submitted;
     }
 
     /**
@@ -656,7 +678,7 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function isValid()
     {
-        if (!$this->bound) {
+        if (!$this->submitted) {
             return false;
         }
 
@@ -724,8 +746,8 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function add($child, $type = null, array $options = array())
     {
-        if ($this->bound) {
-            throw new AlreadyBoundException('You cannot add children to a bound form');
+        if ($this->submitted) {
+            throw new AlreadySubmittedException('You cannot add children to a submitted form');
         }
 
         if (!$this->config->getCompound()) {
@@ -785,8 +807,8 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function remove($name)
     {
-        if ($this->bound) {
-            throw new AlreadyBoundException('You cannot remove children from a bound form');
+        if ($this->submitted) {
+            throw new AlreadySubmittedException('You cannot remove children from a submitted form');
         }
 
         if (isset($this->children[$name])) {
@@ -850,8 +872,8 @@ class Form implements \IteratorAggregate, FormInterface
      * @param string        $name  Ignored. The name of the child is used.
      * @param FormInterface $child The child to be added.
      *
-     * @throws AlreadyBoundException If the form has already been bound.
-     * @throws LogicException        When trying to add a child to a non-compound form.
+     * @throws AlreadySubmittedException If the form has already been submitted.
+     * @throws LogicException            When trying to add a child to a non-compound form.
      *
      * @see self::add()
      */
@@ -865,7 +887,7 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @param string $name The name of the child to remove
      *
-     * @throws AlreadyBoundException If the form has already been bound.
+     * @throws AlreadySubmittedException If the form has already been submitted.
      */
     public function offsetUnset($name)
     {