+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/DateTimeTypeTest.php
@@ -33,7 +33,7 @@ class DateTimeTypeTest extends TypeTestCase
             'input' => 'datetime',
         ));
 
-        $form->bind(array(
+        $form->submit(array(
             'date' => array(
                 'day' => '2',
                 'month' => '6',
@@ -60,7 +60,7 @@ class DateTimeTypeTest extends TypeTestCase
             'time_widget' => 'choice',
         ));
 
-        $form->bind(array(
+        $form->submit(array(
             'date' => array(
                 'day' => '2',
                 'month' => '6',
@@ -85,7 +85,7 @@ class DateTimeTypeTest extends TypeTestCase
             'time_widget' => 'choice',
         ));
 
-        $form->bind(array(
+        $form->submit(array(
             'date' => array(
                 'day' => '2',
                 'month' => '6',
@@ -126,7 +126,7 @@ class DateTimeTypeTest extends TypeTestCase
             ),
         );
 
-        $form->bind($input);
+        $form->submit($input);
 
         $this->assertDateTimeEquals(new \DateTime('2010-06-02 03:00:00 UTC'), $form->getData());
     }
@@ -157,7 +157,7 @@ class DateTimeTypeTest extends TypeTestCase
             ),
         );
 
-        $form->bind($input);
+        $form->submit($input);
 
         $this->assertDateTimeEquals(new \DateTime('2010-06-02 03:04:05 UTC'), $form->getData());
     }
@@ -175,7 +175,7 @@ class DateTimeTypeTest extends TypeTestCase
 
         $dateTime = new \DateTime('2010-06-02 03:04:05 Pacific/Tahiti');
 
-        $form->bind(array(
+        $form->submit(array(
             'date' => array(
                 'day' => (int) $dateTime->format('d'),
                 'month' => (int) $dateTime->format('m'),
@@ -204,7 +204,7 @@ class DateTimeTypeTest extends TypeTestCase
 
         $outputTime = new \DateTime('2010-06-02 03:04:00 Pacific/Tahiti');
 
-        $form->bind('2010-06-02T03:04:00-10:00');
+        $form->submit('2010-06-02T03:04:00-10:00');
 
         $outputTime->setTimezone(new \DateTimeZone('America/New_York'));
 
@@ -221,7 +221,7 @@ class DateTimeTypeTest extends TypeTestCase
             'widget' => 'single_text',
         ));
 
-        $form->bind('2010-06-02T03:04:00Z');
+        $form->submit('2010-06-02T03:04:00Z');
 
         $this->assertEquals('2010-06-02 03:04:00', $form->getData());
         $this->assertEquals('2010-06-02T03:04:00Z', $form->getViewData());
@@ -237,7 +237,7 @@ class DateTimeTypeTest extends TypeTestCase
             'with_seconds' => true,
         ));
 
-        $form->bind('2010-06-02T03:04:05Z');
+        $form->submit('2010-06-02T03:04:05Z');
 
         $this->assertEquals('2010-06-02 03:04:05', $form->getData());
         $this->assertEquals('2010-06-02T03:04:05Z', $form->getViewData());
@@ -254,7 +254,7 @@ class DateTimeTypeTest extends TypeTestCase
 
         $dateTime = new \DateTime('2010-06-02 03:04');
 
-        $form->bind(array(
+        $form->submit(array(
             'date' => '06*2010*02',
             'time' => '03:04',
         ));