+++ src/Symfony/Component/Form/Tests/Extension/HttpFoundation/EventListener/BindRequestListenerTest.php
@@ -84,7 +84,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider requestMethodProvider
      */
-    public function testBindRequest($method)
+    public function testSubmitRequest($method)
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -113,7 +113,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider requestMethodProvider
      */
-    public function testBindRequestWithEmptyName($method)
+    public function testSubmitRequestWithEmptyName($method)
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -140,7 +140,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider requestMethodProvider
      */
-    public function testBindEmptyRequestToCompoundForm($method)
+    public function testSubmitEmptyRequestToCompoundForm($method)
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -167,7 +167,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider requestMethodProvider
      */
-    public function testBindEmptyRequestToSimpleForm($method)
+    public function testSubmitEmptyRequestToSimpleForm($method)
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -190,7 +190,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($event->getData());
     }
 
-    public function testBindGetRequest()
+    public function testSubmitGetRequest()
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -215,7 +215,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
         ), $event->getData());
     }
 
-    public function testBindGetRequestWithEmptyName()
+    public function testSubmitGetRequestWithEmptyName()
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -239,7 +239,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
         ), $event->getData());
     }
 
-    public function testBindEmptyGetRequestToCompoundForm()
+    public function testSubmitEmptyGetRequestToCompoundForm()
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');
@@ -262,7 +262,7 @@ class BindRequestListenerTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(array(), $event->getData());
     }
 
-    public function testBindEmptyGetRequestToSimpleForm()
+    public function testSubmitEmptyGetRequestToSimpleForm()
     {
         if (!class_exists('Symfony\Component\HttpFoundation\Request')) {
             $this->markTestSkipped('The "HttpFoundation" component is not available');