+++ src/Symfony/Component/Form/Tests/Extension/Validator/Constraints/FormValidatorTest.php
@@ -206,7 +206,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->getForm();
 
         // Launch transformer
-        $form->bind(array());
+        $form->submit(array());
 
         $context->expects($this->never())
             ->method('validate');
@@ -235,7 +235,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->getForm();
 
         // Launch transformer
-        $form->bind('foo');
+        $form->submit('foo');
 
         $context->expects($this->never())
             ->method('validate');
@@ -275,7 +275,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->getForm();
 
         // Launch transformer
-        $form->bind('foo');
+        $form->submit('foo');
 
         $context->expects($this->never())
             ->method('validate');
@@ -314,7 +314,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->getForm();
 
         // Launch transformer
-        $form->bind(array());
+        $form->submit(array());
 
         $context->expects($this->never())
             ->method('validate');
@@ -346,7 +346,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->getForm();
 
         // Launch transformer
-        $form->bind(array('child' => 'foo'));
+        $form->submit(array('child' => 'foo'));
 
         $context->expects($this->never())
             ->method('addViolation');
@@ -582,7 +582,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
             ->add($this->getBuilder('child'))
             ->getForm();
 
-        $form->bind(array('foo' => 'bar'));
+        $form->submit(array('foo' => 'bar'));
 
         $context->expects($this->once())
             ->method('addViolation')
@@ -717,7 +717,7 @@ class FormValidatorTest extends \PHPUnit_Framework_TestCase
 
     private function getClickedSubmitButton($name = 'name', array $options = array())
     {
-        return $this->getSubmitButton($name, $options)->bind('');
+        return $this->getSubmitButton($name, $options)->submit('');
     }
 
     /**